requirements.txt
# Core dependencies
faker==19.13.0
sqlalchemy==2.0.25
pandas==2.1.4
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.15.0
joblib==1.3.2
keras==2.15.0

# Web Interface
streamlit==1.29.0
plotly==5.18.0

# API
fastapi==0.109.0
uvicorn==0.27.0

# Data Formats
pyyaml==6.0.1
python-dateutil==2.8.2

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
python-dotenv>=1.0.0
